refactor: 17 review fix, remove BaseIntegrationTest extension from SwitchStepTest, delete get-message.yml, move Condition and SwitchStep to another folder
