--- services/get-message.yml
@@ -1,39 +1,0 @@
-getCurrentTime:
-  assign:
-    currentTime: "Saturday"
-
-conditionalSwitch:
-  switch:
-    - condition: ${currentTime == "Friday"}
-      next: get_message
-    - condition: ${currentTime == "sgsg" || currentTime == "Sunday"}
-      next: post_message
-    - condition: ${currentTime == "dsgsdgsd"}
-      next: return_value
-  next: assign_value
-
-get_message:
-  call: http.get
-  args:
-    url: https://example.com/endpoint
-    query:
-      some_val: "Hello World"
-      another_val: 123
-  result: the_response
-
-post_message:
-  call: http.post
-  args:
-    url: https://example.com/endpoint
-    body:
-      some_val: "Hello World"
-      another_val: 123
-  result: the_message
-
-assign_value:
-  assign:
-    stringValue: "Bürokratt"
-    integerValue: 2021
-
-return_value:
-  return: the_message.body
