+++ src/main/java/ee/buerokratt/ruuter/domain/steps/conditional/SwitchStep.java
@@ -1,8 +1,8 @@
-package ee.buerokratt.ruuter.domain.steps;
+package ee.buerokratt.ruuter.domain.steps.conditional;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
-import ee.buerokratt.ruuter.domain.Condition;
 import ee.buerokratt.ruuter.domain.ConfigurationInstance;
+import ee.buerokratt.ruuter.domain.steps.ConfigurationStep;
 import ee.buerokratt.ruuter.helper.ScriptingHelper;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -25,8 +25,7 @@ public class SwitchStep extends ConfigurationStep {
         super.execute(configurationInstance);
         ScriptingHelper scriptingHelper = configurationInstance.getScriptingHelper();
         Optional<Condition> correctStatement = conditions.stream()
-            .filter(condition -> scriptingHelper.containsScript(condition.getConditionStatement())
-                && Boolean.TRUE.equals(scriptingHelper.evaluateScripts(condition.getConditionStatement(), configurationInstance.getContext())))
+            .filter(condition -> Boolean.TRUE.equals(scriptingHelper.evaluateScripts(condition.getConditionStatement(), configurationInstance.getContext())))
             .findFirst();
         correctStatement.ifPresent(condition -> this.setNextStepName(condition.getNextStepName()));
     }
